requirements.txt
requests~=2.25.1
six~=1.16.0
ImageHash~=4.2.0
Pillow~=8.2.0
